TST: move data to conftest.py
