+++ test/accessor/test_series.py
@@ -1,5 +1,5 @@
-from test.accessor import s
-from test.accessor import s_inf
+from test.accessor.conftest import s
+from test.accessor.conftest import s_inf
 
 import numpy as np
 import pandas as pd
