+++ test/transformer/test_pipeline.py
@@ -1,8 +1,8 @@
-from test.transformer import df_iris
-from test.transformer import df_label
-from test.transformer import df_mixed
-from test.transformer import feature_names
-from test.transformer import s
+from test.transformer.conftest import df_iris
+from test.transformer.conftest import df_label
+from test.transformer.conftest import df_mixed
+from test.transformer.conftest import feature_names
+from test.transformer.conftest import s
 
 import joblib
 import pandas as pd
