+++ test/transformer/test_sklearn.py
@@ -1,6 +1,6 @@
-from test.transformer import df_iris
-from test.transformer import df_label
-from test.transformer import s
+from test.transformer.conftest import df_iris
+from test.transformer.conftest import df_label
+from test.transformer.conftest import s
 
 import pandas as pd
 from scipy import sparse
